pyproject.toml
[build-system]
requires = ["setuptools"]

[project]
name = "public-data-extractor-for-meetup-com"
description = "Public Data Extractor for meetup.com"
readme = "README.md"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests", "PyJWT", "cryptography"
]

[project.optional-dependencies]
dev= ["pytest", "black", "isort", "flake8"]

[tool.isort]
profile = "black"
